:bug: #2788 `unoinId` 使用 Integer 溢出导致GSON抛出NumberFormatException

+++ weixin-java-miniapp/src/main/java/cn/binarywang/wx/miniapp/bean/safety/response/WxMaUserSafetyRiskRankResponse.java
@@ -26,7 +26,7 @@ public class WxMaUserSafetyRiskRankResponse implements Serializable {
    * 唯一请求标识，标记单次请求
    */
   @SerializedName("unoin_id")
-  private Integer unoinId;
+  private Long unoinId;
 
   /**
    * 用户风险等级
